fix: separei as classes acumuladas em Tabuleiro para outros arquivos python
