+++ src/Cadeia.py
@@ -1,0 +1,7 @@
+from .Jogador import Jogador
+from .CasaTabuleiro import CasaTabuleiro
+
+class Cadeia(CasaTabuleiro):
+    def action(self, jogador: Jogador, val_dados: int = 0):
+        pass
+
