+++ src/CasaSorte.py
@@ -1,0 +1,6 @@
+from .Jogador import Jogador
+from .CasaTabuleiro import CasaTabuleiro
+
+class CasaSorte(CasaTabuleiro):
+    def action(self, jogador: Jogador, val_dados: int = 0):
+        pass
