+++ src/EstacionamentoLivre.py
@@ -1,0 +1,6 @@
+from .CasaTabuleiro import CasaTabuleiro
+from .Jogador import Jogador
+
+class EstacionamentoLivre(CasaTabuleiro):
+    def action(self, jogador: Jogador, val_dados: int = 0):
+        pass
