+++ src/PontoDePartida.py
@@ -1,0 +1,7 @@
+from .CasaTabuleiro import CasaTabuleiro
+from .Jogador import Jogador
+
+class PontoDePartida(CasaTabuleiro):
+    def action(self, jogador: Jogador, val_dados: int = 0):
+        pass
+
